fix: the input of  agent chat would be replaced if conversation history updated
