+++ web/features/agent-v2/agent-detail/configure/components/preview/build-chat.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import type { AgentChatRuntimeEmptyStateProps, AgentChatRuntimeProps } from './chat-runtime'
+import type { AgentChatRuntimeProps } from './chat-runtime'
 import { Popover, PopoverContent, PopoverTrigger } from '@langgenius/dify-ui/popover'
 import { useTranslation } from 'react-i18next'
 import { AgentChatRuntime } from './chat-runtime'
@@ -26,66 +26,63 @@ type AgentBuildChatProps = Omit<
   'inputPlaceholder' | 'renderEmptyState' | 'sendButtonLabel'
 >
 
-function AgentBuildChatEmptyState({ inputNode }: AgentChatRuntimeEmptyStateProps) {
+function AgentBuildChatEmptyState() {
   const { t } = useTranslation('agentV2')
   const communityEditionBuildModeTip = t(
     ($) => $['agentDetail.configure.build.empty.communityEditionTip'],
   )
 
   return (
-    <div className="flex h-full items-center justify-center">
-      <div className="flex w-full max-w-150 flex-col items-start p-3 text-left">
-        <div className="dify-blue-glass-surface relative flex h-[50px] w-12 items-center justify-center rounded-xl p-2">
-          <div className="absolute inset-x-px inset-y-0.5 grid grid-cols-[repeat(8,4px)] grid-rows-[repeat(8,4px)] gap-0.5 opacity-25">
-            {buildIconGridCells.map((cell) => (
-              <span
-                key={cell.id}
-                className={cell.opacity > 0 ? 'rounded-[1px] bg-[#98A2B2]' : 'invisible'}
-                style={{ opacity: cell.opacity }}
-              />
-            ))}
-          </div>
-          <span
-            aria-hidden
-            className="absolute i-ri-hammer-line size-5 text-saas-dify-blue-inverted"
-          />
-        </div>
-        <div className="mt-3 flex max-w-full items-center gap-1.5">
-          <div className="min-w-0 truncate system-md-medium text-text-secondary">
-            {t(($) => $['agentDetail.configure.build.empty.title'])}
-          </div>
-          <Popover>
-            <PopoverTrigger
-              openOnHover
-              delay={300}
-              closeDelay={200}
-              aria-label={communityEditionBuildModeTip}
-              render={
-                <button
-                  type="button"
-                  className="inline-flex size-4 shrink-0 items-center justify-center rounded-sm outline-hidden focus-visible:ring-2 focus-visible:ring-state-accent-solid"
-                >
-                  <span
-                    aria-hidden
-                    className="i-custom-vender-line-alertsAndFeedback-alert-triangle size-4 text-text-warning-secondary"
-                  />
-                </button>
-              }
+    <>
+      <div className="dify-blue-glass-surface relative flex h-[50px] w-12 items-center justify-center rounded-xl p-2">
+        <div className="absolute inset-x-px inset-y-0.5 grid grid-cols-[repeat(8,4px)] grid-rows-[repeat(8,4px)] gap-0.5 opacity-25">
+          {buildIconGridCells.map((cell) => (
+            <span
+              key={cell.id}
+              className={cell.opacity > 0 ? 'rounded-[1px] bg-[#98A2B2]' : 'invisible'}
+              style={{ opacity: cell.opacity }}
             />
-            <PopoverContent
-              placement="top"
-              popupClassName="max-w-[340px] px-3 py-2 system-xs-regular text-text-tertiary"
-            >
-              {communityEditionBuildModeTip}
-            </PopoverContent>
-          </Popover>
+          ))}
         </div>
-        <p className="mt-1 max-w-full body-md-regular text-text-tertiary">
-          {t(($) => $['agentDetail.configure.build.empty.description'])}
-        </p>
-        {inputNode}
+        <span
+          aria-hidden
+          className="absolute i-ri-hammer-line size-5 text-saas-dify-blue-inverted"
+        />
       </div>
-    </div>
+      <div className="mt-3 flex max-w-full items-center gap-1.5">
+        <div className="min-w-0 truncate system-md-medium text-text-secondary">
+          {t(($) => $['agentDetail.configure.build.empty.title'])}
+        </div>
+        <Popover>
+          <PopoverTrigger
+            openOnHover
+            delay={300}
+            closeDelay={200}
+            aria-label={communityEditionBuildModeTip}
+            render={
+              <button
+                type="button"
+                className="inline-flex size-4 shrink-0 items-center justify-center rounded-sm outline-hidden focus-visible:ring-2 focus-visible:ring-state-accent-solid"
+              >
+                <span
+                  aria-hidden
+                  className="i-custom-vender-line-alertsAndFeedback-alert-triangle size-4 text-text-warning-secondary"
+                />
+              </button>
+            }
+          />
+          <PopoverContent
+            placement="top"
+            popupClassName="max-w-[340px] px-3 py-2 system-xs-regular text-text-tertiary"
+          >
+            {communityEditionBuildModeTip}
+          </PopoverContent>
+        </Popover>
+      </div>
+      <p className="mt-1 max-w-full body-md-regular text-text-tertiary">
+        {t(($) => $['agentDetail.configure.build.empty.description'])}
+      </p>
+    </>
   )
 }
 
@@ -98,9 +95,7 @@ export function AgentBuildChat(props: AgentBuildChatProps) {
       inputPlaceholder={t(($) => $['agentDetail.configure.build.inputPlaceholder'])}
       inputAutoFocus={false}
       sendButtonLabel={t(($) => $['agentDetail.configure.build.startBuild'])}
-      renderEmptyState={(emptyStateProps: AgentChatRuntimeEmptyStateProps) => (
-        <AgentBuildChatEmptyState {...emptyStateProps} />
-      )}
+      renderEmptyState={() => <AgentBuildChatEmptyState />}
     />
   )
 }
